requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
langgraph==0.0.26
streamlit==1.29.0
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
aiofiles==23.2.1
zipfile36==0.1.3
pathlib==1.0.1
typing-extensions==4.8.0
